+++ build-system/pr-check/experiment-tests.js
@@ -20,13 +20,13 @@
  */
 
 const {
-  getExperimentConfig,
   downloadExperimentOutput,
   printSkipMessage,
   timedExecOrDie,
 } = require('./utils');
 const {buildTargetsInclude, Targets} = require('./build-targets');
 const {experiment} = require('minimist')(process.argv.slice(2));
+const {getExperimentConfig} = require('../common/utils');
 const {runCiJob} = require('./ci-job');
 
 const jobName = `${experiment}-tests.js`;
